requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
torch==2.2.1
torchaudio==2.2.1
numpy==1.23.5
pydub==0.25.1
python-dotenv==1.0.0
